refactor: Move normalization functions to a dedicated module
