+++ packages/cozy-stack-client/src/DocumentCollection.js
@@ -12,6 +12,7 @@ import {
   MangoPartialFilter,
   DesignDoc
 } from './mangoIndex'
+import { normalizeDoc, normalizeDoctypeJsonApi } from './normalize'
 
 import Collection, {
   dontThrowNotFoundError,
@@ -34,39 +35,6 @@ import { FetchError } from './errors'
 import logger from './logger'
 
 const DATABASE_DOES_NOT_EXIST = 'Database does not exist.'
-
-/**
- * Normalize a document, adding its doctype if needed
- *
- * @param {object} doc - Document to normalize
- * @param {string} doctype - Document doctype
- * @returns {object} normalized document
- * @private
- */
-export function normalizeDoc(doc = {}, doctype) {
-  const id = doc._id || doc.id
-  return { id, _id: id, _type: doctype, ...doc }
-}
-
-/**
- * Normalizes a document in JSON API format for a specific doctype
- *
- * @param {string} doctype - The document type
- * @returns {Function} A function that normalizes the document
- */
-export function normalizeDoctypeJsonApi(doctype) {
-  /**
-   * @param {object} data - The document from "data" property of the response in JSON API format
-   * @returns {object} The normalized document
-   */
-  return function(data) {
-    const normalizedDoc = normalizeDoc(data, doctype)
-    return {
-      ...normalizedDoc,
-      ...normalizedDoc.attributes
-    }
-  }
-}
 
 const prepareForDeletion = x =>
   Object.assign({}, omit(x, '_type'), { _deleted: true })
@@ -823,5 +791,3 @@ The returned documents are paginated by the stack.
 }
 
 export default DocumentCollection
-
-export const normalizeDoctype = DocumentCollection.normalizeDoctype
