+++ packages/cozy-stack-client/src/index.js
@@ -1,4 +1,4 @@
 export { default } from './CozyStackClient'
 export { default as OAuthClient } from './OAuthClient'
 export { default as errors, FetchError } from './errors'
-export { normalizeDoc } from './DocumentCollection'
+export { normalizeDoc } from './normalize'
